requirements.txt
requests
python-dotenv
streamlit
openai
pypdf2
typing